requirements.txt
arxiv
feedparser
anthropic
openai
jinja2
pyyaml
pytz
PyExecJS
loguru
requests
